Removed Defunct Functions + ClosestTagMK1
Due to our recent renovation of the tags into a class, a lot of the preexisting functions are no longer necessary. After considerable testing proving that our class does the same things as those functions, we can now safely remove them from the code. Additionally this commit passes an early edition of the next closest tag function, but the functionality of finding all the tags distances from only one is not here yet.