Integrating Tag Class With Nascent Sweet-spot System
In this version, steps have been taken to start communicating with the driver station for certain inputs, and those inputs will tell the robot to go to various sweet-spots, or desired points near the tags. We have started the process of being able to tell where the robot is at anytime with respect to a driver-requested sweet-spot.
Additionally, the depth-rate for the images were reduced, as more memory-intensive vision is not necessary for April-Tags.