Combined Tag Class with New BuildWorld function
Removed previous One_tag_to_all function and updated it to a more effective build word function with deduces the cameras distance from all the tags with only seeing one in frame. Additionally, we changed the camera we're using for the project, so some camera parameters needed to be adjusted a little.