Added Spike support and new Tag Rotation attribute
Rotation: Once you're at a tags sweet spot, you can't tell your orientation since rng = 0. We will still need a way of tracking the robots orientation, and this will be by comparing the way the robot is pointed to a vertical line(Still in the works).

Spikes are new sweetspots for notes at the start of a match, but don't have any april tags associated with them, so further work will have to be done to calculate the distance to them based on any april tags in view.