Added Sweetspot location array
Since we haven't been able to record our ideal sweetspots by the robots location (as we are still waiting for it to be operational), we added some estimated values for the tags (which are just a small offset from the provided values. Additionally, I did the math to figure out the coordinates of the spikes which hold all the notes at the start of the match. These have also been added to that array of sweetspots.