Network tables now display all hotspots
After recently creating the array for the world coordinates of the spikes (locations with notes(game piece) on them without an april tag), they are now also passed into the build world function, updating the bearings table (distance from every hotspot). This commit now updates the network tables to display the distance from every hotspot accessed in the bearings array. Additionally, there was a minor change in jargon for sweet-spot.