Resolved Conflict with altered CSS class
My previous version of the links page worked well until I retrieved an updated version of the file which had substantially altered one of the styles I was using. I made a new one that resembles the previous one and added some notes for things to work on for the next commit.